ref: split previous code to modules, removed old comments
