+++ js/settings.js
@@ -1,0 +1,54 @@
+const AMOUNT_OF_POSTS = 25;
+const DUMMY_COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+const DUMMY_NAMES = [
+  'Аарон', 'Абид', 'Аботур', 'Аввакум', 'Авдей', 'Авигдор', 'Авксентий', 'Авл', 'Авнер', 'Аврелий', 'Автандил', 'Автоном', 'Агапит', 'Агафангел', 'Агафодор', 'Агафон', 'Агриппа', 'Адам', 'Адиль', 'Адольф', 'Адонирам', 'Адриан', 'Азамат', 'Азат', 'Азиз', 'Азим', 'Айварс', 'Айдар', 'Акакий', 'Аквилий', 'Акиндин', 'Акиф', 'Акоп', 'Аксель', 'Алан', 'Аланус', 'Александр', 'Алексей', 'Алик', 'Алим', 'Алипий', 'Алишер', 'Алоиз', 'Альберик', 'Альберт', 'Альбин', 'Альваро', 'Альвизе', 'Альфонс', 'Альфред', 'Амадис', 'Амвросий', 'Амедей', 'Амин', 'Амир', 'Амр', 'Анания', 'Анас', 'Анастасий', 'Анатолий', 'Андокид', 'Андреан', 'Андреас', 'Андрей', 'Андроник', 'Аникий', 'Аникита', 'Аннерс', 'Анри', 'Ансельм', 'Антипа', 'Антон', 'Антоний', 'Антонин', 'Арефа', 'Арзуман', 'Аристарх', 'Ариф', 'Аркадий', 'Арсен', 'Арсений', 'Артём', 'Артемий', 'Артур', 'Арфаксад', 'Архипп', 'Атанасий', 'Аттик', 'Афанасий', 'Афинагор', 'Афиней', 'Африкан', 'Ахилл', 'Ахмад', 'Ахтям', 'Ашот', 'Бадр', 'Барни', 'Басир', 'Бахтияр', 'Бен', 'Бехруз', 'Билял', 'Богдан', 'Болеслав', 'Болот', 'Бонавентура', 'Борис', 'Борислав', 'Боян', 'Бронислав', 'Брячислав', 'Булат', 'Бурхан', 'Бямбасурэн', 'Вадим', 'Валентин', 'Валерий', 'Валерьян', 'Вальдемар', 'Вангьял', 'Варлам', 'Варнава', 'Варсонофий', 'Варфоломей', 'Василий', 'Вахтанг', 'Велвел', 'Велимир', 'Вениамин', 'Венцеслав', 'Верослав', 'Викентий', 'Виктор', 'Викторин', 'Вильгельм', 'Винцас', 'Виссарион', 'Виталий', 'Витаутас', 'Владимир', 'Владислав', 'Владлен', 'Влас', 'Волк', 'Володарь', 'Вольфганг', 'Вописк', 'Всеволод', 'Всеслав', 'Вук', 'Вукол', 'Вячеслав', 'Гавриил', 'Гай', 'Галактион', 'Гарет', 'Гаспар', 'Гафур', 'Гвидо', 'Гейдар', 'Геласий', 'Гельмут', 'Геннадий', 'Генри', 'Генрих', 'Георге', 'Георгий', 'Гераклид', 'Герберт', 'Герман', 'Германн', 'Геронтий', 'Герхард', 'Гессий', 'Гильем', 'Гинкмар', 'Глеб', 'Гней', 'Гонорий', 'Горацио', 'Гордей', 'Гостомысл', 'Гоффредо', 'Градислав', 'Григорий', 'Гримоальд', 'Груди', 'Гуго', 'Гьялцен', 'Давид', 'Дамдинсурэн', 'Дамир', 'Данакт', 'Даниил', 'Дарий', 'Демид', 'Демьян', 'Денис'
+];
+const DUMMY_SURNAMES = [
+  'Иванов', 'Петров', 'Сидоров', 'Смирнов', 'Кузнецов', 'Попов', 'Васильев', 'Соколов', 'Михайлов', 'Новиков', 'Фёдоров', 'Морозов', 'Волков', 'Алексеев', 'Лебедев', 'Семенов', 'Егоров', 'Павлов', 'Козлов', 'Степанов', 'Николаев', 'Орлов', 'Белоусов', 'Крылов', 'Карпов', 'Воробьёв', 'Денисов', 'Богданов', 'Поляков', 'Анисимов', 'Тарасов', 'Горшков', 'Захаров', 'Калинин', 'Куликов', 'Артемьев', 'Герасимов', 'Никитин', 'Григорьев', 'Лазарев', 'Медведев', 'Кораблёв', 'Голубев', 'Овчинников', 'Кудрявцев', 'Прохоров', 'Наумов', 'Панфилов', 'Кириллов', 'Маркелов', 'Соболев', 'Громов', 'Виноградов', 'Фомин', 'Дементьев', 'Мельников', 'Щербаков', 'Блинов', 'Колесников', 'Кондратьев', 'Баранов', 'Щукин', 'Фролов', 'Артамонов', 'Зубов', 'Суворов', 'Тимофеев', 'Орехов', 'Ермаков', 'Дроздов', 'Игнатьев', 'Савельев', 'Логинов', 'Сафонов', 'Капустин', 'Кирсанов', 'Маслов', 'Родионов', 'Гурьев', 'Елисеев', 'Ефремов', 'Исаев', 'Евдокимов', 'Калашников', 'Кабанов', 'Носков', 'Юдин', 'Кулагин', 'Лапин', 'Комаров', 'Александров', 'Соловьёв', 'Рыбаков', 'Кочергин', 'Вавилов', 'Гончаров', 'Прокофьев', 'Филатов', 'Королёв', 'Климов', 'Пантелеев', 'Устинов', 'Миронов', 'Кудряшов', 'Белов', 'Гаврилов', 'Ермолов', 'Мартынов', 'Лазарев', 'Одинцов', 'Ежов', 'Никонов', 'Гаврилин', 'Золотарёв', 'Воронцов', 'Чистяков', 'Симонов', 'Кожевников', 'Оськин', 'Буторин'
+];
+const DUMMY_DESCRIPTIONS = {
+  which: [
+    'красный', 'синий', 'зелёный', 'жёлтый', 'чёрный', 'белый', 'серый', 'фиолетовый', 'оранжевый', 'розовый', 'голубой', 'коричневый', 'золотой', 'серебряный'
+  ],
+  who: [
+    'кот', 'дом', 'человек', 'мост', 'город', 'цветок', 'корабль', 'лес', 'парк', 'пляж', 'пёс'
+  ],
+  what: [
+    'бежит', 'сидит', 'летит', 'прыгает', 'стоит', 'плавает', 'играет', 'смотрит', 'спит', 'думает', 'работает', 'разговаривает', 'рисует', 'поёт', 'читаёт', 'улыбается', 'катается', 'гуляет', 'строит', 'танцует'
+  ],
+  where: [
+    'в лесу', 'на пляже', 'в городе', 'на озере', 'в парке', 'на улице', 'в горах', 'в поле', 'на крыше дома', 'на мосту', 'на берегу реки', 'в саду', 'на площади', 'в замке', 'на стадионе', 'в пустыне', 'в аэропорту'
+  ],
+  when: [
+    'утром', 'днём', 'вечером', 'ночью', 'на рассвете', 'на закате', 'при лунном свете', 'во время дождя', 'при тумане'
+  ],
+  hashtags: [
+    '#nature', '#photography', '#travel', '#art', '#beautiful', '#instagood', '#photooftheday', '#love', '#fashion',
+    '#summer', '#sunset', '#naturelovers', '#landscape', '#adventure', '#happy', '#fun', '#smile', '#beach', '#friends',
+    '#food', '#fitness', '#style', '#motivation', '#beauty', '#design', '#inspiration', '#music', '#life', '#workout',
+    '#health', '#mindfulness', '#urban', '#city', '#mountains', '#sky', '#sea', '#photo', '#picoftheday', '#family',
+    '#creative', '#quotes', '#flowers', '#sun', '#dog', '#cat', '#animals', '#selfie', '#lifestyle', '#happiness',
+    '#goodvibes', '#travelgram', '#explore', '#photographer', '#naturephotography', '#loveit', '#instamood', '#fitnessmotivation',
+    '#wanderlust', '#skylovers', '#instatravel', '#architecture', '#minimal', '#night', '#party', '#weekend', '#relax',
+    '#hiking', '#camping', '#wildlife', '#streetphotography', '#retro', '#memories', '#blackandwhite', '#ocean', '#water',
+    '#cool', '#igers', '#follow', '#instadaily', '#repost', '#daily', '#best', '#artwork', '#dream', '#mindset', '#positivevibes',
+    '#weekendvibes', '#winter', '#snow', '#coffee', '#tea', '#reading', '#writing', '#film', '#cinema', '#portrait',
+    '#moments', '#colorful', '#calm', '#exploremore'
+  ]
+};
+
+
+export {
+  AMOUNT_OF_POSTS,
+  DUMMY_COMMENTS,
+  DUMMY_NAMES,
+  DUMMY_SURNAMES,
+  DUMMY_DESCRIPTIONS
+}
